Rewrite static file handling
 - Let freemarker handle inclusion of static files
 - Put static resources in public dir
 - Don't fetch 'prettify' async, include it with other scripts
 - Auto format static resources
